python-ml-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
scikit-learn==1.3.2
numpy==1.26.2
python-multipart==0.0.6
transformers
torch
torchvision
google-generativeai
Pillow
